perf(sqlite): enable WAL, busy_timeout and NORMAL sync on every connection
Every operation opens its own SQLite connection with default rollback
journaling and no busy_timeout, so a UI read colliding with a
background job write (import, face detection, thumbnails) failed
immediately with SQLITE_BUSY. WAL allows concurrent reads during
writes, busy_timeout=5000 makes remaining collisions wait, and
synchronous=NORMAL drops the per-commit fsync that rollback journaling
forced on batch operations.

Connection management (struct, init, get_connection) moves from
sqlite/mod.rs to sqlite/connection.rs to respect the file-size limit;
no behavior change beyond the new PRAGMAs, covered by a regression
test.
